renaming register
