+++ src/pages/register/index.ts
@@ -1,1 +1,1 @@
-export { RegisterPage } from './register'
+export { RegisterPage } from './tmp_Register'
